requirements.txt
blinker
click
colorama
Flask
greenlet
gunicorn
h11
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
protobuf
PyMySQL
python-dotenv
SQLAlchemy
typing_extensions
uvicorn
waitress
Werkzeug
